ford angle mode: gain model moves to values_ext — strategy owns it, calibrator consumes (review 5.2)
GAIN_CAN / GAIN_CANFD_BOF / GAIN_CANFD_SUV, the per-platform CAR sets,
platform_gains(), and the V_LOW / V_HIGH / LOW_ANCHOR_BASE anchors now live in
values_ext.py. Kills the inverted dependency (strategy re-importing its own
gain model out of the calibrator through a noqa'd second import block) and the
inline duplicate of platform_gains() in update_angle_params. The strategy's
gain interp now uses the named anchors instead of hardcoded 13.5/26.82/1.30 —
one source for the numbers the calibrator's fit is expressed against.

Pure refactor: 37 tests pass, route 0000000b analyzer replay byte-identical.
